22. Личный проект: модель для сборки
