+++ .eslintignore
@@ -1,2 +1,2 @@
 webpack.config.js
-js/debounce.js
+js/bundle.js
