+++ .gitignore
@@ -11,3 +11,4 @@ npm-debug.log*
 Thumbs.db
 .vscode/
 js/mock.js
+js/bundle.js
