+++ index.html
@@ -248,15 +248,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
-  <script src="js/util.js"></script>
-  <script src="js/debounce.js"></script>
-  <script src="js/message.js"></script>
-  <script src="js/beckend.js"></script>
-  <script src="js/pin.js"></script>
-  <script src="js/card.js"></script>
-  <script src="js/form.js"></script>
-  <script src="js/filter.js"></script>
-  <script src="js/draganddrop.js"></script>
+  <script src="js/bundle.js"></script>
+
 </body>
 </html>
