--- js/debounce.js
@@ -1,18 +1,0 @@
-'use strict';
-(function () {
-  let DEBOUNCE_INTERVAL = 500; // ms
-
-  window.debounce = function (cb) {
-    let lastTimeout = null;
-
-    return function () {
-      let parameters = arguments;
-      if (lastTimeout) {
-        window.clearTimeout(lastTimeout);
-      }
-      lastTimeout = window.setTimeout(function () {
-        cb.apply(null, parameters);
-      }, DEBOUNCE_INTERVAL);
-    };
-  };
-})();
