+++ js/modules/debounce.js
@@ -1,0 +1,16 @@
+'use strict';
+
+const DEBOUNCE_INTERVAL = 500; // ms
+
+window.debounce = function (cb) {
+  let lastTimeout = null;
+
+  return function (...parameters) {
+    if (lastTimeout) {
+      window.clearTimeout(lastTimeout);
+    }
+    lastTimeout = window.setTimeout(function () {
+      cb(...parameters);
+    }, DEBOUNCE_INTERVAL);
+  };
+};
